Extract scene functions to their respective modules
- Also, extract utilities into utils module -- capitalizeString, getDateTime, isInputLike
