+++ src/scene/createScene.ts
@@ -1,0 +1,6 @@
+import { ExcalidrawElement } from "../element/types";
+
+export const createScene = () => {
+  const elements = Array.of<ExcalidrawElement>();
+  return { elements };
+};
